Actually switch functions based on mode instead of storing them in storage
While doing this (and during testing) I noticed more annotation issues and fixed them.
